deploy/requirements.txt
# Core Libraries
fastapi==0.110.2
uvicorn[standard]==0.29.0

# PyTorch ecosystem (CUDA 12.1 builds)
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Hugging Face Transformers & Dependencies
transformers==4.41.1
tokenizers==0.19.1
sentencepiece==0.2.0
# Utilities
numpy==1.26.4
pyyaml==6.0.1
python-multipart==0.0.9
jinja2==3.1.3
soundfile==0.13.1
